refactor: move methods to Sections
