+++ src/main/java/nextstep/subway/domain/model/Line.java
@@ -1,11 +1,9 @@
 package nextstep.subway.domain.model;
 
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
@@ -13,16 +11,10 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import org.springframework.data.util.Pair;
-
 import nextstep.subway.domain.service.SectionAdditionStrategy;
 
 @Entity
 public class Line {
-    public static final String  FIRST_SECTION_NOT_FOUND_MESSAGE = "첫번째 구간을 찾을 수 없습니다.";
-
-    public static final String  LAST_SECTION_NOT_FOUND_MESSAGE = "마지막 구간을 찾을 수 없습니다.";
-
     public static final String SECTION_NOT_FOUND_MESSAGE = "구간을 찾을 수 없습니다.";
 
     @Id
@@ -88,61 +80,8 @@ public class Line {
         sections.removeSection(station);
     }
 
-    private Section getFirstSection() {
-        return sections.getSections().stream()
-            .filter(section -> sections.getSections().stream()
-                .noneMatch(s -> s.getDownStation().equals(section.getUpStation())))
-            .findFirst()
-            .orElseThrow(() -> new IllegalStateException(FIRST_SECTION_NOT_FOUND_MESSAGE));
-    }
-
-    public Section getLastSection() {
-        return sections.getSections().stream()
-            .filter(section -> sections.getSections().stream()
-                .noneMatch(s -> s.getUpStation().equals(section.getDownStation())))
-            .findFirst()
-            .orElseThrow(() -> new IllegalStateException(LAST_SECTION_NOT_FOUND_MESSAGE));
-    }
-
-    public List<Station> getOrderedUnmodifiableStations() {
-        if (sections.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        Pair<Map<Station, Section>, Map<Station, Section>> stationToSectionMaps = getStationToSectionMaps();
-        Section firstSection = getFirstSection(stationToSectionMaps.getSecond());
-        List<Station> orderedStations = getOrderedStations(firstSection, stationToSectionMaps.getFirst());
-        return Collections.unmodifiableList(orderedStations);
-    }
-
-    private Pair<Map<Station, Section>, Map<Station, Section>> getStationToSectionMaps() {
-        Map<Station, Section> upStationToSectionMap = new HashMap<>();
-        Map<Station, Section> downStationToSectionMap = new HashMap<>();
-
-        for (Section section : sections.getSections()) {
-            upStationToSectionMap.put(section.getUpStation(), section);
-            downStationToSectionMap.put(section.getDownStation(), section);
-        }
-
-        return Pair.of(upStationToSectionMap, downStationToSectionMap);
-    }
-
-    private List<Station> getOrderedStations(Section firstSection, Map<Station, Section> upStationToSectionMap) {
-        List<Station> orderedStations = new ArrayList<>();
-        Section currentSection = firstSection;
-
-        while (currentSection != null) {
-            orderedStations.add(currentSection.getUpStation());
-            Section nextSection = upStationToSectionMap.get(currentSection.getDownStation());
-
-            if (nextSection == null) {
-                orderedStations.add(currentSection.getDownStation());
-            }
-
-            currentSection = nextSection;
-        }
-
-        return orderedStations;
+    public Optional<Section> getLastSection() {
+        return sections.getLastSection();
     }
 
     public List<Section> getSections() {
@@ -154,36 +93,11 @@ public class Line {
     }
 
     public List<Section> getOrderedUnmodifiableSections() {
-        if (sections.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        Pair<Map<Station, Section>, Map<Station, Section>> stationToSectionMaps = getStationToSectionMaps();
-        Section firstSection = getFirstSection(stationToSectionMaps.getSecond());
-        List<Section> orderedSections = getOrderedSections(firstSection, stationToSectionMaps);
-        return Collections.unmodifiableList(orderedSections);
+        return sections.getOrderedUnmodifiableSections();
     }
 
-    private List<Section> getOrderedSections(
-        Section firstSection,
-        Pair<Map<Station, Section>, Map<Station, Section>> stationToSectionMaps
-    ) {
-        List<Section> orderedSections = new ArrayList<>();
-        Section currentSection = firstSection;
-
-        while (currentSection != null) {
-            orderedSections.add(currentSection);
-            currentSection = stationToSectionMaps.getFirst().get(currentSection.getDownStation());
-        }
-        return orderedSections;
-    }
-
-    private Section getFirstSection(Map<Station, Section> downStationToSectionMap) {
-        return sections.getSections()
-            .stream()
-            .filter(section -> !downStationToSectionMap.containsKey(section.getUpStation()))
-            .findFirst()
-            .orElse(null);
+    public List<Station> getOrderedUnmodifiableStations() {
+        return sections.getOrderedUnmodifiableStations();
     }
 
     @Override
