+++ src/main/java/nextstep/subway/domain/model/Sections.java
@@ -4,11 +4,16 @@ import static nextstep.subway.domain.model.Line.*;
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
+
+import org.springframework.data.util.Pair;
 
 import nextstep.subway.domain.service.SectionAdditionStrategy;
 
@@ -24,6 +29,95 @@ public class Sections {
     private final List<Section> sections = new ArrayList<>();
 
     protected Sections() {
+    }
+
+    public List<Station> getOrderedUnmodifiableStations() {
+        if (sections.isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        Pair<Map<Station, Section>, Map<Station, Section>> stationToSectionMaps = getStationToSectionMaps();
+        Section firstSection = getFirstSection(stationToSectionMaps.getSecond());
+        List<Station> orderedStations = getOrderedStations(firstSection, stationToSectionMaps.getFirst());
+        return Collections.unmodifiableList(orderedStations);
+    }
+
+    private List<Station> getOrderedStations(Section firstSection, Map<Station, Section> upStationToSectionMap) {
+        List<Station> orderedStations = new ArrayList<>();
+        Section currentSection = firstSection;
+
+        while (currentSection != null) {
+            orderedStations.add(currentSection.getUpStation());
+            Section nextSection = upStationToSectionMap.get(currentSection.getDownStation());
+
+            if (nextSection == null) {
+                orderedStations.add(currentSection.getDownStation());
+            }
+
+            currentSection = nextSection;
+        }
+
+        return orderedStations;
+    }
+
+    public List<Section> getOrderedUnmodifiableSections() {
+        if (sections.isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        Pair<Map<Station, Section>, Map<Station, Section>> stationToSectionMaps = getStationToSectionMaps();
+        Section firstSection = getFirstSection(stationToSectionMaps.getSecond());
+        List<Section> orderedSections = getOrderedSections(firstSection, stationToSectionMaps);
+        return Collections.unmodifiableList(orderedSections);
+    }
+
+    private Pair<Map<Station, Section>, Map<Station, Section>> getStationToSectionMaps() {
+        Map<Station, Section> upStationToSectionMap = new HashMap<>();
+        Map<Station, Section> downStationToSectionMap = new HashMap<>();
+
+        for (Section section : sections) {
+            upStationToSectionMap.put(section.getUpStation(), section);
+            downStationToSectionMap.put(section.getDownStation(), section);
+        }
+
+        return Pair.of(upStationToSectionMap, downStationToSectionMap);
+    }
+
+    private Section getFirstSection(Map<Station, Section> downStationToSectionMap) {
+        return sections
+            .stream()
+            .filter(section -> !downStationToSectionMap.containsKey(section.getUpStation()))
+            .findFirst()
+            .orElse(null);
+    }
+
+    public Optional<Section> getLastSection() {
+        if (sections.isEmpty()) {
+            return Optional.empty();
+        }
+
+        return getLastSection(getStationToSectionMaps().getFirst());
+    }
+
+    private Optional<Section> getLastSection(Map<Station, Section> upStationToSectionMap) {
+        return sections
+            .stream()
+            .filter(section -> !upStationToSectionMap.containsKey(section.getDownStation()))
+            .findFirst();
+    }
+
+    private List<Section> getOrderedSections(
+        Section firstSection,
+        Pair<Map<Station, Section>, Map<Station, Section>> stationToSectionMaps
+    ) {
+        List<Section> orderedSections = new ArrayList<>();
+        Section currentSection = firstSection;
+
+        while (currentSection != null) {
+            orderedSections.add(currentSection);
+            currentSection = stationToSectionMaps.getFirst().get(currentSection.getDownStation());
+        }
+        return orderedSections;
     }
 
     public void addSection(Line line, Section newSection) {
